Change JSON results format
To make it easier to handle multiple runtimes and to make it easier to use the returned data in the UI, the JSON format of the results is changed from using patterns like `[["name", {..}], ...]` to `{"name": {...}, ...}`. To coordinate the format between sightglass and sg-history, `test/fixtures` contains example JSON used in the newly-created unit tests. The history metadata is also expanded to include the Git references to multiple runtimes.