Alter UI to display multiple runtimes
This is a large change in how the UI works; it makes use of the new data format to chart multiple lines on the graph (one for each runtime). The filtering and re-construction of objects in this change may not be optimal for extremely large history sizes but that performance tuning is left for a future commit.