Port runner and webui branches to master
Ports the runner and webui branches to master. The runner code
has been put under a new webui_runner folder and includes the
docker support, the python driver and a pluggable infrastructure
for downloading and benchmarking the latest VM updates. An expected
workflow is to launch the webui using the provided docker scripts
and to separately run an experiment that captures these results
for immediate viewing and additionally sends the results to the
webui server (expected at localhost by default) for storing and
displaying history data.