Integrate a new test runner
This change brings in work done in a separate repository to create a new benchmark runner for sightglass. The original code is retained with a `legacy` moniker where appropriate; the new runner uses `next`. The `README.md` now contains a description of the different components present in the repository and links to documentation or implementation for each. In writing this, it clarified for me that a reorganization (in a future PR) is needed.

The new runner employs a different paradigm than the old one:
 - benchmarks are Wasm files, not shared libraries, built using Dockerfiles and committed in-tree (accompanied by their WAT pair) under `benchmarks-next`; they   import `bench.start` and `bench.end` functions which are placed around the section of code to measure
 - engines (i.e. Wasm VMs) are shared libraries, like the one introduced in https://github.com/bytecodealliance/wasmtime/pull/2437, allowing the runner to control the compilation, instantiation, and execution of the Wasm benchmark
 - the building and running of the benchmarks is controlled with a CLI tool, see `cargo +nightly run -- --help` and `docs/next.md` for more information