Add support for taking samples from multiple processes and engines
This provides the base infrastructure that will allow us to compare "control"
and "variable" engines (e.g. the latest `main` build and a build from a branch
with a potential optimization), recording metrics from both engines across
multiple Wasm programs.

For example, to record data for two versions of Wasmtime's performance on
`{a,b,c,d}.wasm` you can run:

```
$ cargo run -p sightglass-cli -- \
    benchmark \
    -e libwasmtime.control.so \
    -e libwasmtime.variable.so \
    a.wasm b.wasm c.wasm d.wasm
```

The runner will shuffle the process executions so that we don't run all of the
control processes before all of the variable processes, which could accidentally
inflict measurement bias on our results due to, for example, dynamic CPU
throttling.