Update for new benchmarking API
This means that we can now pass in file paths where stdout/stderr should be
logged to, an (optional) file path that can be read and passed to Wasm execution
as its stdin, and we can instantiate the module multiple times per
compilation (although this feature is not hooked up to the CLI yet).

The management of starting and ending timers is also simplified with the new
benchmark API, and so we have no more need for the `static_measure` module and
its trickery.