Add a measurement mechanism for capturing phases in VTune
This adds a `--measure vtune` option to the `benchmark` command that
informs the VTune engine of the various benchmarking phases
(compilation, instantiation, execution). This new measurement option
does not output any data (like the `noop` measure); instead, VTune
receives "task" markers that indicate when each phase starts and ends.
This facilitates performance analysis by allowing us to focus in on the
phase we're interested in.

This change still has one major TODO:
 - the `wasmtime-bench-api` crate does not dump its JIT-compiled code,
   which is necessary for VTune to inspect the code in question; this
   functionality exists (see `ProfilingStrategy` in the [VTune docs])
   but we must find a way to configure this from Sightglass

[VTune docs]: https://docs.wasmtime.dev/examples-profiling-vtune.html#turn-on-vtune-support