Add SpiderMonkey benchmark.
This benchmark builds SpiderMonkey using its support for running on
`wasm32-wasi` platforms, and links it with a simple toplevel that feeds
it a recursive Fibonacci function and evaluates `fib(33)`.

On my system, this benchmark has the following timing measurements:

```
compilation
  benchmarks-next/spidermonkey/benchmark.wasm
    cycles
      [8740808933 9440410499.20 9781710940] ./wasmtime.so
    nanoseconds
      [3807417356 4112157473.80 4260824860] ./wasmtime.so
instantiation
  benchmarks-next/spidermonkey/benchmark.wasm
    cycles
      [271009 288346.40 303439] ./wasmtime.so
    nanoseconds
      [118049 125600.80 132175] ./wasmtime.so
execution
  benchmarks-next/spidermonkey/benchmark.wasm
    cycles
      [7055231738 7076764356.40 7114914898] ./wasmtime.so
    nanoseconds
      [3073195167 3082574580.80 3099192612] ./wasmtime.so
```

It is a good test of both compile-time (SpiderMonkey itself is a large
program and takes the engine a nontrivial amount of time to compile) and
runtime (the interpreter loop is CPU-bound and involves several large
functions that are stress-tests of generated-code quality).