Add build info to Sightglass engines
When building an engine for Sightglass, this change will capture key
details about how it was built for later reproduction of the build. When
comparing benchmarks across different machines, it is critical that the
WebAssembly engine used be the same, or as close to the same, as
possible. The build information captured by this change (stored in
`.build-info` files) is an approximation of [reproducible builds]; it is
only an approximation since there are certain environmental factors that
might need to change (e.g., OS, architecture).

[reproducible builds]: https://reproducible-builds.org/docs/definition/