recorder: stub out a mechanism for passing execution flags to engines (#169)
In https://github.com/bytecodealliance/wasmtime/pull/4096, I added a way
to configure an engine using Wasmtime's CLI flags. This change adds some
stub fields in Sightglass for this flag string but does not fully wire
it up. I have a solution for configuring the execution flags using
`BuildInfo` in #167 so I will fully wire this up there; in the meantime,
Sightglass should still work as before with this stub implementation.