Build Wasmtime engine using a `build.rs` script
The current mechanism for building engines using Docker has several
issues: it is slow (Cargo must re-download all required crates), it is
too Linux-specific, and it does not play well with workflows that
attempt to build Sightglass in Docker (i.e., docker-in-docker is
tricky). This change provides a `build.rs` script to build both the
engine shared library and a `.build-info` metadata file: the engine
shared library is built using the same steps the `Dockerfile` used and
the `.build-info` file attempts to capture some of the environmental
information that Docker fixed but is now system-dependent (e.g., tool
versions for reproducing the build).