Remove CLI commands--`build-engine`, `clean`
If we no longer build Sightglass engines using Docker (see previous
commit) and store them in a central place (the Sightglass cache
directory), then it no longer makes sense to have the `build-engine` and
`clean` commands available in the CLI. This change removes these
commands and associated dead code.

Alternately, we could remove Docker but not the cache directory; by
default, the Wasmtime engine `build.rs` script could place built files
there. The difficulty with this is that we previously used the `dirs`
crate to reliably find the `data_dir()` within the user's home
directory. It is difficult to use crates from a standalone script like
`build.rs` and without a reliable way to access the user's home
directory, it seemed easiest to just copy the built files to
`engines/wasmtime`.