Log the current processor number (#171)
While benchmarking, it can be quite important to know what core a
benchmark is assigned to: it could be useful to detect scheduling
conflicts, e.g., but more importantly, on hybrid architectures the
choice of core drastically affects the performance. On Alder Lake cores,
e.g., the performance cores (0-15) are significantly faster than the
efficiency cores (16-23) and I have noticed large variance in benchmark
results when the core affinity is not set.

Fortunately, core affinity can be set; on Linux, `taskset --cpu-list
0-15 cargo run -- benchmark ...` will limit the benchmark execution to
the "right" set of cores. In order to check this, I used this change;
though it is possible that benchmarks could be re-scheduled mid-workload
to a different core, this logging gave me some confidence that, at least
this point, the cores were scheduled correctly. I weighed the downside
of adding a dependency to `libc` with the benefit of knowing which core
the benchmark iteration is running on and thought this change was
worthwhile--if logging is turned off there should be no effect but when
you are looking for this information it is quite useful.

(Note: the current mechanism for determining the processor ID is
Linux-specific; if anyone finds an OS-agnostic way to do this, please
update this!)