Build engines without Docker (#170)
* Build Wasmtime engine using a `build.rs` script

The current mechanism for building engines using Docker has several
issues: it is slow (Cargo must re-download all required crates), it is
too Linux-specific, and it does not play well with workflows that
attempt to build Sightglass in Docker (i.e., docker-in-docker is
tricky). This change provides a `build.rs` script to build both the
engine shared library and a `.build-info` metadata file: the engine
shared library is built using the same steps the `Dockerfile` used and
the `.build-info` file attempts to capture some of the environmental
information that Docker fixed but is now system-dependent (e.g., tool
versions for reproducing the build).

* Remove CLI commands--`build-engine`, `clean`

If we no longer build Sightglass engines using Docker (see previous
commit) and store them in a central place (the Sightglass cache
directory), then it no longer makes sense to have the `build-engine` and
`clean` commands available in the CLI. This change removes these
commands and associated dead code.

Alternately, we could remove Docker but not the cache directory; by
default, the Wasmtime engine `build.rs` script could place built files
there. The difficulty with this is that we previously used the `dirs`
crate to reliably find the `data_dir()` within the user's home
directory. It is difficult to use crates from a standalone script like
`build.rs` and without a reliable way to access the user's home
directory, it seemed easiest to just copy the built files to
`engines/wasmtime`.

* ci: use the `build.rs` script

Instead of using the `build-engine` command (now removed) or manually
building the engine library, use the `build.rs` script throughout the CI
workflow files.

* fix: check that alternate engine exists