ui: remove old UI
In expectation that we will replace the existing UI with something else,
this change removes the old UI code and the associated helper scripts
that kept it fed with measurement data.

Some background: the old UI was created by @jedisct1 for the original
version of sightglass. It used a Rust HTTP backend that maintained a
single JSON file as its database. The frontend, a Vue.js application,
was modified over time to adapt to the changing definition of a
benchmark "measurement;" however, in recent times it has not been
actively maintained and it is unclear how much work it would be to fix
it. Later, @jlb6740 added a harness for periodically running the
benchmarks and feeding the results to this Rust/Vue.js system.

Since the "single JSON file" database was limited (no built-in
filtering, slower lookups over time) and since other off-the-shelf
solutions do exist, I recommend we abandon the maintenance of this code
and remove it from the repository. This should reduce the maintenance
burden for this repository (e.g., see the number of outdated JS
dependency PRs that have to be audited and merged).

Where do we go from here? I have a working proof-of-concept of an
Elasticsearch database with a Kibana UI that runs from a
`docker-compose` command. My changes add a `sightglass-cli upload`
command to send results to this database. The primary issue to resolve
there is an investigation of ES's licensing requirements--does their new
licensing make that a prohibitive solution? If that does not work out,
we should be able to switch to a different database/UI solution with
minor changes to the `sightglass-cli upload` code.