Port the ability to pin to a single core from the legacy runner
In the legacy runner, @jedisct1 added the ability to pin execution to a
single core, using several different mechanisms: it uses the `core_affinity`
crate for common OSes and the `hwloc` crate otherwise. This change ports
that functionality over to the new runner under the `--pin` flag.

This is the last feature that existed in the legacy runner but not in
the new runner (that I could see, at least). Though core-pinning can be
accomplished otherwise, e.g, `taskset --cpu-list 0`, these mechanisms
apply to all subprocesses. It could be possible that a user wants to
benchmark using multiple processes, all on potentially different CPUs,
but have all iterations in that process be pinned to a single core to
avoid measuring context switches.