Organize benchmarks into suites
This change closes #202 by adding a way to group benchmarks into
`*.suite` files. This incorporates @jameysharp's comments, e.g., to
calculate the benchmark path relative to the suite path and to detect
suite files based on their `.suite` extension.

Though more suites could be added in the future, the result of this is
that suite files can be built (see `benchmarks/shootout.suite`, e.g.)
and then run like:

```console
$ sightglass-cli benchmark benchmarks/shootout.suite -e ...
```