Allow for different benchmark .wasm names
Changes to build.sh that allow benchmarks to have
a different name than benchmark.wasm. This allows
them to opt out of ./run-all.sh if they choose.
Users can provide the expected .wasm name if they
want. If one isn't provided, it will look for the
filename in the benchmark being built.