ci: add option to run workflows manually (#223)
* ci: add option to run workflows manually

This results in a "Run workflow" button on the Actions page. See the
docs for [`workflow_dispatch`].

[`workflow_dispatch`]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_dispatch

* fingerprint: upgrade `sysinfo` library

This change upgrades the `sysinfo` library to `0.28` to avoid
compilation issues on Windows due to an unaligned packed field (e.g.,
this [failure]). `sysinfo`'s API has changed slightly and this brings
Sightglass' usage up to speed.

[failure]: https://github.com/bytecodealliance/sightglass/actions/runs/4089163019/jobs/7051580568#step:6:459