Add a regular expression benchmark
This regular expression benchmark measures the time taken to match
e-mails, URIs, and IP addresses in an input text. It is written in Rust
and significantly adapted from the [regex-benchmark] project. The input
text is a bit heavy (6.6MB) and the `benchmark.wasm` is as well (2.8MB);
it may be worthwhile at some point to run some of these benchmark files
through `wasm-opt`, e.g., to reduce the size.

It appears that one iteration of this benchmark takes around half a
second for all three phases and the compilation phase is about an order
of magnitude larger than the execution phase. I looked into creating a
smaller version of the input text, but this would not significantly
change the overall benchmark time (since it is concentrated in
compilation) and, importantly, I could not figure out a way to have
different `stdout` and `stderr` expectations for different workloads.
With a smaller workload, the output _will_ be different and currently
there is no way to express that.

[regex-benchmark]: https://github.com/mariomka/regex-benchmark