Build Rust benchmarks from a common `Dockerfile`
Emscripten- and wasi-sdk-built projects already use a common
`Dockerfile` for rebuilding the benchmarks. This change extends that to
the Rust-built benchmarks by symlink-ing to `Dockerfile.rust`. The idea
is to reduce duplication and make it easier to change build
configuration in one place. For example, this change also bumps the used
version of Rust to 1.70, which causes the emitted WebAssembly to
slightly change.