Update `hex-simd` expected output
Previously, the `hex-simd` benchmark retrieved the bytes to hexify from
its own WebAssembly binary. This self-referential use was prone to
breakage any time the `benchmark.wasm` changed, like from building with
a new version of Rust. The `stdout.expected` file would immediately be
out of date. To fix this, this change randomly dumps 1MB of data into
`default.input` (as other benchmarks do) and uses this as the input to
hexify.