Run each `libsodium` benchmark for one iteration
Previously, we compiled in the default number of iterations to run--200.
This made the execution times for `libsodium` much too long to run
during CI. Since Sightglass already has a mechanism for controlling the
number of runs, `--processes` and `--iterations-per-process`, this
commit compiles the `libsodium` benchmarks to only run a single
iteration internally.