Add `simd.suite` (#268)
As discussed in a Cranelift meeting, it would be nice to know which
benchmarks are SIMD users. This change adds a `simd.suite` file that
lists the files that contain SIMD instructions. This can be run like:

```console
$ cargo run -- benchmark --engine .../libengine.so benchmarks/simd.suite
```

There are a couple ways this could be improved, but this may be fine
as-is for now:
- `simd.suite` could get out of date; someone could create a CI check to
  keep it in-sync with the added files
- someone could really check that the measured section of each benchmark
  is actually a heavy SIMD user by collecting the instruction mix (e.g.,
  `sde`)