Add `report` subcommand to generate html report
This is a direct port of the viz.py proof-of-concept.  The
workflow for generate these reports is to:

1. Capture raw data to JSON/CSV.
2. Generate a report using 2 or more input data sets with
   one engine/flags designated as the baseline (by default
   the first engine found in the provided data files).

The report itself tries to provide clear, concise representation
of the data that is statistically significant in the overview
table with raw results data in detail sections which, along with
statistics, can give an indication of how noisy and trustworthy
differences in measured results for a run might be.

The report is intended to be shared; it does pull in a couple
dependencies for graph generation from a CDN but is otherewise
standalone.