Record and Use Engine Flags in Summarize/Effect-Size Tools (#293)
* Allow benchmark names to be overriden

This can be useful when the engine filename is not useful
or when wanting a concise way to differentiate benchmark
runs that may use the same engine but with different flags.

* Record engine_flags with measurement data

Previously, only the engine name/path was stored for each measurement.
The flags used represents valuable information for both understanding
how a historical run was performed as well as determining how different
configurations impact performance for the same engine.

* Update summarize/effect-size to use engine flags

Previously, recordings were updated to capture
engine flag information.  This change updates the code
used for summaries and effect-size analysis to allow
for comparing engine/flag combinations against each other
which is useful for a class of comparisons.

* Fix issues raised by the tests, improve coverage

* Use single field for sightglass_data::Engine

This simplifies consuming and building measurement and other
related structures but at the cost of somewhat more complex
serialization.  This is largely due to our use of CSV as
an input/output format and rust-csv's longstanding inability
to handle flattened structures.

See https://github.com/BurntSushi/rust-csv/issues/98