Bump version of libloading to 0.9
We're seeing spurious crashes on windows after successful benchmark runs
when unloading the dll. Upgrade libloading to the latest as a first step
in diagnosing potential problems.

This updated version may just fix the issue; if not, it provides a way
to do an explicit close() where we can handle errors potentially.